refactor(afmt): separate source formatting from batch adapter
